Remove warning
Changing this to type int removes a warning flag that both sides are
different type. size_t was used in class and recommended by the book,
but I will stick with what VS recommends.

+++ Game.cpp
@@ -127,7 +127,7 @@ void Game::getPlayers() {
         std::cin.ignore();
     } while ((numPlayers < 1) || (numPlayers > 4));
 
-    for (size_t i{1}; i <= numPlayers; ++i) {
+    for (int i{1}; i <= numPlayers; ++i) {
         std::cout << "\nWhat is player " << i << "'s name?: ";
         std::cin >> name;
         do {
